# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API dokumentace aplikace",
        "contact": {
            "name": "API Support",
            "email": "info@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.example.com/v1",
            "description": "Test server"
        }
    ],
    "paths": {
        "/api/authors": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Seznam všech autorů",
                "description": "Vrátí seznam všech autorů s možností filtrování, řazení a stránkování",
                "operationId": "getAuthorsList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Číslo stránky",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Počet položek na stránku",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Sloupec pro řazení",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "surname",
                            "enum": [
                                "name",
                                "surname",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "order",
                        "in": "query",
                        "description": "Směr řazení",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Filtrovat podle jména",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "surname",
                        "in": "query",
                        "description": "Filtrovat podle příjmení",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Úspěšná operace",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Authors"
                ],
                "summary": "Vytvoření nového autora",
                "description": "Vytvoří nového autora a vrátí jeho data",
                "operationId": "storeAuthor",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "surname"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "J.K."
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "Rowlingová"
                                    },
                                    "birthdate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1965-07-31"
                                    },
                                    "biography": {
                                        "type": "string",
                                        "example": "Britská spisovatelka..."
                                    },
                                    "nationality": {
                                        "type": "string",
                                        "example": "britská"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Autor byl úspěšně vytvořen",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validační chyba"
                    }
                }
            }
        },
        "/api/authors/{id}": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Detail autora",
                "description": "Vrátí detailní informace o autorovi podle ID",
                "operationId": "getAuthorById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID autora",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Úspěšná operace",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Autor nebyl nalezen"
                    }
                }
            },
            "put": {
                "tags": [
                    "Authors"
                ],
                "summary": "Aktualizace autora",
                "description": "Aktualizuje autora podle ID a vrátí jeho data",
                "operationId": "updateAuthor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID autora",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "J.K."
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "Rowling"
                                    },
                                    "birthdate": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "1965-07-31"
                                    },
                                    "biography": {
                                        "type": "string",
                                        "example": "Aktualizovaná biografie..."
                                    },
                                    "nationality": {
                                        "type": "string",
                                        "example": "britská"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Autor byl úspěšně aktualizován",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Autor nebyl nalezen"
                    },
                    "422": {
                        "description": "Validační chyba"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Authors"
                ],
                "summary": "Smazání autora",
                "description": "Smaže autora podle ID",
                "operationId": "deleteAuthor",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID autora",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Autor byl úspěšně smazán"
                    },
                    "404": {
                        "description": "Autor nebyl nalezen"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}